pyproject.toml
[tool.poetry]
name = "streamlit-agent"
version = "0.1.0"
description = ""
authors = ["Tim Conkling <tconkling@gmail.com>", "Joshua Carroll <joshua.carroll@snowflake.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "streamlit_agent"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
langchain = ">=0.0.216"
streamlit = ">=1.24.0"
openai = "^0.27.8"
duckduckgo-search = "^3.8.3"
watchdog = "^3.0.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.4.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
